pyproject.toml
[project]
name = "Inventory-Agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "openai>=2.7.2",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]
